Bring over bootstrapping logic from builtools
Bring over the bootstrapping logic from dotnet/buildtools so we can get
a CLI to use.

The boostrapping logic in dotnet/buildtools still assumes a project.json
based CLI and so it generates a project.json to restore. The preview 3
RTW tooling no longer supports project.json so this fails.

Sicne we don't need to bring down any additional tools at this point (we
should be raw MSBuild), I've just removed the logic that handles
project.json generation and restoration of the packages listed in
.toolversions.