Update Core-Setup Sources
This pulls in a set of fixes to the core-setup build system.

The major changes are using --infer-runtimes in more places (so we don't
have to explicitly list or restore runtimes we don't care about for some
projects), teaching update-dependencies to both work on *nix and to
allow pointing it at local files for the version info instead of just
remote files and having the Microsoft.NETCore.App build use the root
NuGet.config instead of hard-coding it's own sources.