Use Microsoft.DotNet.PlatformAbstractions to compute RID
Unfortuantly, using PackageReference to
Microsoft.DotNet.PlatformAbstractions doesn't work because at runtime
GetHostRid can't load that libray (likely due to MSBuild not
understanding where to resolve the assembly from).

It looks like this issue may be fixed in newer versions of MSBuild,
but we might need to move to publishing the tasks so that
M.D.PlatformAbstractions is SxS with M.D.S.Tasks. Until we get that
all figured out, vendor in parts of M.D.PlatformAbstractions.