Don't allow TargetRid to leak across targets
Since MSBuild will merge all properties and items defined in a target into
the global enviromnet, we were running into an issue where TargetRid was
getting defined to the Host RID and then not overridden in future targets
which need to control the RID we use (e.g. to always use win7-x64 when
building packages on Windows).

Restructure our properties such that things with well known names don't
leak out of targets.

Long term we should figure out a better strategy here (I think the right
idea is that we undefine all properties and item groups before a target
ends with the exception of a few things we want to leak for composition
reasons (e.g. the ItemGroup that has all of the version info we will use
for update-dependencies).