Refactor how host specific information is computed
Previously, we assumed all non windows builds were on Linux, which was
not correct for OSX. In addition, each task would compute a value for
TargetRid if it needed it.

Add a feature to our task that gets host information to also get a
"name" for the current OS (Linux, Windows_NT, OSX) which matches the
name that buildtools based repositories use for these OSes and have an
initial target that runs and set some properties based on the current
OS for use in other targets.