Docker images for building
This changes adds Dockerfiles that describe a build environment for
.NET Core for CentOS 7, Ubuntu 14.0 and Fedora Core 23.

In addition there are some helper scripts to automate building an
image from a Dockerfile and then spin up a transient docker container
based on that image, with the source code mounted in it. The scripts
do work to ensure that the UID of the worker *inside* the container
matches the UID of the user that invoked the script *outside* the
container so any generated files are owned by the user instead of
root.

Example usage:

```
./scripts/docker-run.sh -d docker/ubuntu.14.04/Dockerfile ./build.sh
```

This will build an image based on the
docker/ubuntu.14.04/Dockerfile (this will be slow the first time but
docker caching on a local machine will make everything faster on the
next invocation, the spin up a container with the source code mounted
in /code (which will be the CWD for the command). When the command
finishes, the docker container will be removed.  Artifacts from the
build continue to be present in `<repo-root>/bin` as expected.

At some point I would like to move CI to consume these docker
images. We'll either need to push images to an Azure docker
registry (the right long term thing) or dockerhub (a good short term
hack).