Bump Submodules
As part of this change, I updated the URL for the submodules for all
projects except for libuv and coreclr to be pointing at dotnet instead
of ellismg. libuv remains pointed at ellismg for now since it's the
upstream location for the prototype and we're using ellismg for
coreclr right now so we can move faster without having to wait for
changes to land in master.