Remove Libuv and bump submodules
LibUV is no longer going to be shipped as part of the shared
framework. Instead, folks that don't want to have to ship libuv as
part of their application will use the new `dotnet-cache` mechanism to
cache it on the machine they wish to deploy to and then use policy at
publish time to not include it as part of the published app.

Because of this, we can just remove all of libuv from our repository.