Introduce RepositoryReference support
Previously build ordering was defined via phases.props and we manually
controlled the order in which things built by putting them in
different phases.

This change moves us to a model where each repository lists the other
repositories that it consumes resources from, via the
RepositoryReference item group.

When building a repository, we first build all parallel
repositories. We depend on the fact that MSBuild will only build a
single project once if all the global properties are the same.

I've added repository references in a few places (cli, core-setup,
corefx, roslyn) where I know we have dependencies and we care about
the build order. We'll augment additional projects as time moves on.

For now, if you do not add a repository reference to CLI (or something
it depends on), your repository will not be built.