Enable building all configurations in corefx package build
Corefx only builds packages since the last servcing in release
branches so it will be disabled for most packages and they
will come from nuget.org instead when restoring.

To get the package build correct we also need to stop passing
in the OfficialBuildId unless this is actually an official build.
Once we start doing official orchestrated builds it can be past
through. Most of the properties being passed aren't interesting
for live git builds only for tar-ball type builds they are now
conditioned on OfflineBuild