Add new OS bootstrap doc and script
This change adds a document on creating a bootstrap cli for a new
unsupported OS and a script that automates the whole process.
While there is still space for improvements in both the doc and the
script, I think it is time to publish them. I have used them to bring up
RHEL 6 and Alpine and tested it on FreeBSD for x64. I have also tested
it on arm64 Linux.