Change default bootstrap type to release
Bootstrapping works with a seed CLI. The CLI distributed by Microsoft is
a release build. The CLI built by source-build is a release build.
There's a much greater chance that the default cli will be a release
build rather than a debug build. An incompatible configuration will not
work. So change the default from debug to release to make it more likely
to work.

Also update the docs to point out that the seed CLI configuration must
match the configuration being used here or the dotnet command will fail
at runtime with cryptic error messages.