Update source-build master to core repo masters (#658)
Most of the work here was in adding the new toolset and core-sdk repos and adjusting our current process to the way they work. I also had the usual build number/hash updates. This includes a few workarounds and patches that I believe are all tagged with an associated issue or PR.