Do not run dotnet when bootstrapping coreclr
This makes it easier to bootstrap incompatible Linux x86_64
distributions. The existing builds of .NET Core won't run
there but the boostrap script will still (indirectly) try and run
`dotnet` when building coreclr. That `dotnet` invocation will fail
causing the bootstrap process to fail.

This depends on the coreclr submodule being updated to pull in
https://github.com/dotnet/coreclr/pull/19111.

Fixes #663