Build all corefx packages (#653)
* Build all corefx packages

CoreFX includes a number of packages that are not built when using
source-build. For example, Microsoft.NETCore.Platforms is not built. It
is consumed via NuGet. This causes two issues:

1. The bits included in the final cli build are not all compiled from
source. This goes against overall goal of source-build.

3. Any changes made to such packages in corefx do not flow through. As
an example, this means additional runtime IDs added to
Microsoft.NETCore.Platforms are not included in the final built cli.

Fix that by passing a flag to indicate all of CoreFX should be built.

Fixes #650

* Add separate folder for previously source-built packages that are used earlier in the build than they are produced.

* Update the prebuilt baseline.