Split build and pack for aspnet-razor project
If we only do a "pack", the items are resolved before the build happens. Razor has some direct includes of files in "bin", which are missed because compilation hasn't happened yet. Splitting to individual build/pack means that the files are all built when "pack" project evaluation searches for them.