Update init-tools to handle additional architectures
This is based on the platfrm selection logic from coreclr.

Instead of trying to run x64 executibles on non-x64 platforms, this
forces the download of the platform-appropriate dotnetcli and then uses
that.

This commit also doesn't actually fix builds on any platform. It just
gets buildtools running on additional platforms. Depending on the
platform, things will still fail.