Enable building on arm64 (#1300)
* Enable building on arm64

This change allows source-build to be cloned and built on arm64
machines. It shouldn't affect cross compilation at all, however. I have
tested this on RHEL 8 aarch64.

This change includes backports of the following pull-requests which have
been merged into their respective master branches:

- https://github.com/aspnet/AspNetCore/pull/14790
- https://github.com/aspnet/AspNetCore/pull/15354
- https://github.com/dotnet/core-sdk/pull/4102
- https://github.com/dotnet/core-setup/pull/8468
- https://github.com/dotnet/corefx/pull/40453

There's a number of existing build configuration that are
conditionalized on arm64, such as setting up a root file system. Those
they are actually only meant to be invoked when cross-compiling for
arm64 (on x86_64). This commit modifies those conditions to not apply
when building on an arm64 machine.

* Patch fixup.