Fix task dev in Visual Studio when rolled forward
There were two breaks here: 1: the SDK was being rolled forward to one installed globally (and with .NET Framework MSBuild there don't seem to be any easy properties to navigate to the global SDK's DLL dir), 2: with the change to use the Microsoft.NET.Sdk more directly, the NETCoreSdkVersion is not yet defined in the Directory.Build.props chain. The simplest way to fix both is to allow any locally-installed SDK to work, regardless of what SDK is being used to build.