Update build arg for sdk repo and remove installer patch.
The sdk repo is now building with a newer version of the sdk
in source-build than it is using in the official build.  The new
version defaults to producing reference assemblies.  Because of this
there was a tools/5.0/ref/ folder being generated which caused issues
with crossgen in installer.  The patch was filtering this new
ref assembly out.  Instead, don't build reference assemblies
as the sdk repo expects and remove the patch.