Add dotnet/cssparser repo and build it from source
This repo produces the Microsoft.Css.Parser nuget package which is used
by aspnetcore.

dotnet/cssparser is licensed under MIT. It does not use arcade, so the
build project is based off of application-insights.proj.

After this fix, a local tarball build results in the removal of this
line from the annotated prebuilt xml:

    <AnnotatedUsage Id="Microsoft.Css.Parser"
        Version="1.0.0-20200708.1"
        File="src/aspnetcore.a4938d07a5127ffad8466ddf703a6b5b21f4b0c9/artifacts/obj/Microsoft.AspNetCore.Razor.Tools/project.assets.json"
        IsDirectDependency="true"
        IsAutoReferenced="true"
        Project="src/aspnetcore.a4938d07a5127ffad8466ddf703a6b5b21f4b0c9/"/>