Fix smoke test timeout detection to work again
.NET Core 2.1 and 3.1 use separate strings of text to indicate that the
application is running. Both include this text, though:

    Application started. Press Ctrl+C to shut down.

Update the timeout detection logic to be more flexible and catch this
line of text anywhere in the output, not just in the last line.

This brings the timeout for some of these test runs down from 30 seconds
to 2 or 3 seconds.

This is a port of #1963 to master.