Reuse SDK directory for version parsing and smoke-tests
Modify smoke-test.sh to use the same extracted cli directory as the one
we use to parse the SDK/Runtime/ASP version.

The smoke-test.sh script offers to interactively delete
$SCRIPT_ROOT/smoke-testing if it exists. That interactive prompt breaks
CI. So we can't use the smoke-testing dir. Instead, point smoke-test.sh
to a different directory (outside of $SCRIPT_ROOT/smoke-testing) so it
can reuse that.